Completed Family/Group new application.

+++ backend/src/routes/application.js
@@ -13,10 +13,10 @@ const {
   getApprovableFeeSchedules,
   getDeclaration,
   submitApplication,
-  updateStatusForApplicationComponents,
 } = require('../components/application');
 const { getNMFApplication, updateNMFApplication, createNMFApplication } = require('../components/nmfApplication');
-const { UUID_VALIDATOR_VERSION } = require('../util/constants');
+const validatePermission = require('../middlewares/validatePermission');
+const { PERMISSIONS, UUID_VALIDATOR_VERSION } = require('../util/constants');
 const { param, validationResult, body } = require('express-validator');
 
 router.post('/renew-ccof', passport.authenticate('jwt', { session: false }), isValidBackendToken, [], (req, res) => {
@@ -93,8 +93,7 @@ router.delete(
   },
 );
 
-router.patch('/ccfri', passport.authenticate('jwt', { session: false }), isValidBackendToken, [], (req, res) => {
-  //validationResult(req).throw();
+router.patch('/ccfri', passport.authenticate('jwt', { session: false }), isValidBackendToken, validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION), (req, res) => {
   return updateCCFRIApplication(req, res);
 });
 
@@ -155,8 +154,7 @@ router.put(
 /* CREATE or UPDATE parent fees for a specified age group and year.
   age group and year are defined in the payload
 */
-router.patch('/parentfee', passport.authenticate('jwt', { session: false }), isValidBackendToken, [], (req, res) => {
-  //validationResult(req).throw();
+router.patch('/parentfee', passport.authenticate('jwt', { session: false }), isValidBackendToken, validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION + 's'), [], (req, res) => {
   return upsertParentFees(req, res);
 });
 
@@ -165,17 +163,19 @@ router.get(
   '/ecewe/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION, PERMISSIONS.VIEW_SUBMITTED_PCF),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return getECEWEApplication(req, res);
   },
 );
 
-/* Update an ECEWE applciation for an application id. */
+/* Update an ECEWE application for an application id. */
 router.patch(
   '/ecewe/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return updateECEWEApplication(req, res);
@@ -187,6 +187,7 @@ router.post(
   '/ecewe/facilities/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return updateECEWEFacilityApplication(req, res);
@@ -198,6 +199,7 @@ router.get(
   '/declaration/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION, PERMISSIONS.VIEW_SUBMITTED_PCF),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return getDeclaration(req, res);
@@ -209,20 +211,10 @@ router.patch(
   '/declaration/submit/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.SUBMIT_NEW_APPLICATION),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return submitApplication(req, res);
-  },
-);
-
-/* Get the full summary of the application */
-router.get(
-  '/summary/:applicationId',
-  passport.authenticate('jwt', { session: false }),
-  isValidBackendToken,
-  [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
-  (req, res) => {
-    return getApplicationSummary(req, res);
   },
 );
 
@@ -231,29 +223,19 @@ router.post(
   '/summary/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION, PERMISSIONS.VIEW_SUBMITTED_PCF),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION), body('facilities').isArray(), body('facilities.*').isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return getApplicationSummary(req, res);
   },
 );
 
-router.put(
-  '/status/:applicationId',
-  passport.authenticate('jwt', { session: false }),
-  isValidBackendToken,
-  [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
-  (req, res) => {
-    validationResult(req).throw();
-    return updateStatusForApplicationComponents(req, res);
-  },
-);
-
-/*   Get existing change requests for an application */
-
+/* Get existing change requests for an application */
 router.get(
   '/changeRequest/:applicationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  // TODO #securitymatrix Add with Change Requests security
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return getChangeRequest(req, res);
@@ -261,11 +243,11 @@ router.get(
 );
 
 /* DELETE an existing PCF -- new PCF ONLY */
-
 router.delete(
   '/:applicationId/',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
+  validatePermission(PERMISSIONS.CREATE_NEW_APPLICATION),
   [param('applicationId', 'URL param: [applicationId] is required').notEmpty().isUUID(UUID_VALIDATOR_VERSION)],
   (req, res) => {
     return deletePcfApplication(req, res);
